Move IpFamily to config
